refactor: abstract date-time range picker into reusable component
- Create DateTimeRangePicker component in frontend/components/ui/
- Move all picker logic, state, and UI from calendar-grid-view
- Component handles date selection, time pickers, and duration logic
- Maintains Google Calendar-style UX with calendar widget and dropdowns
- Clean props interface for easy reuse across the application
- Remove duplicate code and improve maintainability
